Refactored DiskDataset code to be better packaged.
