+++ api/model/dataset.go
@@ -16,6 +16,9 @@
 package model
 
 import (
+	"fmt"
+	"path"
+
 	"github.com/pkg/errors"
 	log "github.com/unchartedsoftware/plog"
 
@@ -23,6 +26,7 @@ import (
 	"github.com/uncharted-distil/distil-compute/model"
 	"github.com/uncharted-distil/distil-compute/primitive/compute"
 	"github.com/uncharted-distil/distil/api/env"
+	"github.com/uncharted-distil/distil/api/serialization"
 	"github.com/uncharted-distil/distil/api/util/json"
 )
 
@@ -35,16 +39,6 @@ const (
 	// DatasetTypeInference is a dataset consumed by a model to infer predictions.
 	DatasetTypeInference DatasetType = "inference"
 )
-
-// RawDataset contains basic information about the structure of the dataset as well
-// as the raw learning data.
-type RawDataset struct {
-	ID              string
-	Name            string
-	Metadata        *model.Metadata
-	Data            [][]string
-	DefinitiveTypes bool
-}
 
 // Dataset represents a decsription of a dataset.
 type Dataset struct {
@@ -93,6 +87,167 @@ type VariableUpdate struct {
 	Index string `json:"index"`
 	Name  string `json:"name"`
 	Value string `json:"value"`
+}
+
+// DiskDataset represents a dataset stored on disk.
+type DiskDataset struct {
+	Dataset           *serialization.RawDataset
+	FeaturizedDataset *DiskDataset
+	schemaPath        string
+}
+
+// LoadDiskDataset loads a dataset from disk.
+func LoadDiskDataset(ds *Dataset) (*DiskDataset, error) {
+	folder := env.ResolvePath(ds.Source, ds.Folder)
+	output, err := loadDiskDatasetFromFolder(folder)
+	if err != nil {
+		return nil, err
+	}
+
+	if ds.LearningDataset != "" {
+		pre, err := loadDiskDatasetFromFolder(ds.LearningDataset)
+		if err != nil {
+			return nil, err
+		}
+		output.FeaturizedDataset = pre
+	}
+
+	return output, nil
+}
+
+func loadDiskDatasetFromFolder(folder string) (*DiskDataset, error) {
+	schemaPath := path.Join(folder, compute.D3MDataSchema)
+	dsDisk, err := serialization.ReadDataset(schemaPath)
+	if err != nil {
+		return nil, err
+	}
+
+	return &DiskDataset{
+		Dataset:    dsDisk,
+		schemaPath: schemaPath,
+	}, nil
+}
+
+// SaveDataset saves a dataset to disk.
+func (d *DiskDataset) SaveDataset() error {
+	err := serialization.WriteDataset(path.Dir(d.schemaPath), d.Dataset)
+	if err != nil {
+		return err
+	}
+
+	if d.FeaturizedDataset != nil {
+		err = d.FeaturizedDataset.SaveDataset()
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// AddField adds a field to the dataset.
+func (d *DiskDataset) AddField(variable *model.Variable) error {
+	err := d.Dataset.AddField(variable)
+	if err != nil {
+		return err
+	}
+
+	if d.FeaturizedDataset != nil {
+		err = d.FeaturizedDataset.AddField(variable)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// FieldExists returns true if a field exists in the dataset.
+func (d *DiskDataset) FieldExists(variable *model.Variable) bool {
+	return d.Dataset.FieldExists(variable)
+}
+
+// Clone clones a dataset on disk.
+func (d *DiskDataset) Clone(targetFolder string, cloneDatasetID string, cloneStorageName string) (*DiskDataset, error) {
+	// easiest clone is to write the dataset to the new location then read it
+	err := serialization.WriteDataset(targetFolder, d.Dataset)
+	if err != nil {
+		return nil, err
+	}
+
+	schemaPath := path.Join(targetFolder, compute.D3MDataSchema)
+	cloned, err := serialization.ReadDataset(schemaPath)
+	if err != nil {
+		return nil, err
+	}
+
+	// update the metadata
+	cloned.Metadata.ID = cloneDatasetID
+	cloned.Metadata.StorageName = cloneStorageName
+	err = serialization.WriteMetadata(schemaPath, cloned.Metadata)
+	if err != nil {
+		return nil, err
+	}
+
+	output := &DiskDataset{
+		Dataset:    cloned,
+		schemaPath: schemaPath,
+	}
+
+	if d.FeaturizedDataset != nil {
+		clonedPrefeaturized, err := d.FeaturizedDataset.Clone(fmt.Sprintf("%s-featurized", targetFolder), cloneDatasetID, cloneStorageName)
+		if err != nil {
+			return nil, err
+		}
+		output.FeaturizedDataset = clonedPrefeaturized
+	}
+
+	return output, nil
+}
+
+// GetLearningFolder returns the folder containing the learning dataset.
+func (d *DiskDataset) GetLearningFolder() string {
+	if d.FeaturizedDataset != nil {
+		return d.FeaturizedDataset.GetLearningFolder()
+	}
+
+	return path.Dir(d.schemaPath)
+}
+
+// UpdateDataset updates a dataset, optionally filtering rows not being updated.
+func (d *DiskDataset) UpdateDataset(updates map[string]map[string]string, filterNotFound bool) error {
+	if filterNotFound {
+		// do an initial filter pass to keep only the rows found in the updates
+		filterMap := map[string]bool{}
+		for _, colUpdates := range updates {
+			for key := range colUpdates {
+				filterMap[key] = true
+			}
+		}
+		d.Dataset.FilterDataset(filterMap)
+	}
+
+	// translate column names to indices
+	updatesMapped := map[int]map[string]string{}
+	for colName, colUpdates := range updates {
+		index := d.Dataset.GetVariableIndex(colName)
+
+		// could add missing columns!
+		if index == -1 {
+			return errors.Errorf("column %s not in dataset for updates", colName)
+		}
+		updatesMapped[index] = colUpdates
+	}
+	d.Dataset.UpdateDataset(updatesMapped)
+
+	if d.FeaturizedDataset != nil {
+		err := d.FeaturizedDataset.UpdateDataset(updates, filterNotFound)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 // ParseVariableUpdateList returns a list of parsed variable updates.
@@ -195,101 +350,6 @@ func (d *Dataset) GetLearningFolder() string {
 		return d.LearningDataset
 	}
 	return env.ResolvePath(d.Source, d.Folder)
-}
-
-// SyncMetadata updates the key metadata properties to match a given metadata.
-// This is often use to update the metadata for prediction or prefeaturization purposes.
-func (d *RawDataset) SyncMetadata(metaToSync *model.Metadata) {
-	d.Metadata.ID = metaToSync.ID
-	d.Metadata.Name = metaToSync.Name
-	d.Metadata.StorageName = metaToSync.StorageName
-}
-
-// AddField adds a field to the dataset, updating both the data and the metadata.
-func (d *RawDataset) AddField(variable *model.Variable) error {
-	if d.FieldExists(variable) {
-		return errors.Errorf("field '%s' already exists in the raw dataset", variable.Key)
-	}
-	clone := variable.Clone()
-	clone.Index = len(d.Metadata.GetMainDataResource().Variables)
-	d.Metadata.GetMainDataResource().Variables = append(d.Metadata.GetMainDataResource().Variables, clone)
-
-	// the first row is the header row
-	d.Data[0] = append(d.Data[0], variable.HeaderName)
-	for i, row := range d.Data[1:] {
-		d.Data[i+1] = append(row, "")
-	}
-
-	return nil
-}
-
-// FieldExists returns true if a field is already part of the metadata.
-func (d *RawDataset) FieldExists(variable *model.Variable) bool {
-	for _, v := range d.Metadata.GetMainDataResource().Variables {
-		if v.Key == variable.Key {
-			return true
-		}
-	}
-
-	return false
-}
-
-// GetVariableIndex returns the index of the variable as found in the header
-// or -1 if not found in the header.
-func (d *RawDataset) GetVariableIndex(variableHeaderName string) int {
-	for i, f := range d.Data[0] {
-		if f == variableHeaderName {
-			return i
-		}
-	}
-
-	return -1
-}
-
-// GetVariableIndices returns the mapping of variable header name to header index.
-// It will error if a field is not found in the header.
-func (d *RawDataset) GetVariableIndices(variableHeaderNames []string) (map[string]int, error) {
-	indices := map[string]int{}
-	for _, v := range variableHeaderNames {
-		varIndex := d.GetVariableIndex(v)
-		if varIndex == -1 {
-			return nil, errors.Errorf("variable '%s' does not exist in header", v)
-		}
-		indices[v] = varIndex
-	}
-
-	return indices, nil
-}
-
-// FilterDataset updates the dataset to only keep the rows that have the specified
-// column in the filter map set to true.
-func (d *RawDataset) FilterDataset(filter map[string]bool) {
-	d3mIndexIndex := d.GetVariableIndex(model.D3MIndexFieldName)
-
-	// start with the header
-	filteredData := [][]string{d.Data[0]}
-	for i := 1; i < len(d.Data); i++ {
-		if filter[d.Data[i][d3mIndexIndex]] {
-			filteredData = append(filteredData, d.Data[i])
-		}
-	}
-	d.Data = filteredData
-}
-
-// UpdateDataset updates a dataset with the value specified in the updates dictionary.
-// If the specified column value is not found in the dictionary, then it is left unchanged.
-// Updates are specified by column index value.
-func (d *RawDataset) UpdateDataset(updates map[int]map[string]string) {
-	d3mIndexIndex := d.GetVariableIndex(model.D3MIndexFieldName)
-	for i := 1; i < len(d.Data); i++ {
-		d3mIndexValue := d.Data[i][d3mIndexIndex]
-		for columnIndex, colUpdates := range updates {
-			updateValue, ok := colUpdates[d3mIndexValue]
-			if ok {
-				d.Data[i][columnIndex] = updateValue
-			}
-		}
-	}
 }
 
 // UpdateExtremas updates the variable extremas based on the data stored.
