+++ api/serialization/csv.go
@@ -29,7 +29,6 @@ import (
 	"github.com/uncharted-distil/distil-compute/metadata"
 	"github.com/uncharted-distil/distil-compute/model"
 	"github.com/uncharted-distil/distil-compute/primitive/compute"
-	api "github.com/uncharted-distil/distil/api/model"
 	"github.com/uncharted-distil/distil/api/util"
 	log "github.com/unchartedsoftware/plog"
 )
@@ -45,7 +44,7 @@ func NewCSV() *CSV {
 
 // ReadDataset reads a raw dataset from the file system, loading the csv
 // data into memory.
-func (d *CSV) ReadDataset(schemaFile string) (*api.RawDataset, error) {
+func (d *CSV) ReadDataset(schemaFile string) (*RawDataset, error) {
 	meta, err := d.ReadMetadata(schemaFile)
 	if err != nil {
 		return nil, err
@@ -56,7 +55,7 @@ func (d *CSV) ReadDataset(schemaFile string) (*api.RawDataset, error) {
 		return nil, err
 	}
 
-	return &api.RawDataset{
+	return &RawDataset{
 		ID:       meta.ID,
 		Name:     meta.Name,
 		Data:     data,
@@ -66,7 +65,7 @@ func (d *CSV) ReadDataset(schemaFile string) (*api.RawDataset, error) {
 
 // WriteDataset writes the raw dataset to the file system, writing out
 // the data to a csv file.
-func (d *CSV) WriteDataset(uri string, data *api.RawDataset) error {
+func (d *CSV) WriteDataset(uri string, data *RawDataset) error {
 
 	dataFilename := path.Join(uri, compute.D3MDataFolder, compute.D3MLearningData)
 	err := d.WriteData(dataFilename, data.Data)
